move bin tests in test/bin folder
